Add docs `Datetime::class`.
