+++ docs/README.md
@@ -51,6 +51,7 @@ This is particularly useful for creating forms and other interactive elements on
 
 - [date](/docs/form/input/Date.md)
 - [datetime-local](/docs/form/input/DatetimeLocal.md)
+- [datetime](/docs/form/input/Datetime.md)
 - [text](/docs/form/input/Text.md)
 - [time](/docs/form/input/Time.md)
 - [url](/docs/form/input/Url.md)
