+++ tests/Input/Datetime/AttributeTest.php
@@ -143,26 +143,6 @@ final class AttributeTest extends TestCase
         );
     }
 
-    public function testMax(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="datetime-6582f2d099e8b" type="datetime" max="value">
-            HTML,
-            Datetime::widget()->id('datetime-6582f2d099e8b')->max('value')->render()
-        );
-    }
-
-    public function testMin(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="datetime-6582f2d099e8b" type="datetime" min="value">
-            HTML,
-            Datetime::widget()->id('datetime-6582f2d099e8b')->min('value')->render()
-        );
-    }
-
     public function testLang(): void
     {
         Assert::equalsWithoutLE(
@@ -190,26 +170,6 @@ final class AttributeTest extends TestCase
             <input id="datetime-6582f2d099e8b" type="datetime" readonly>
             HTML,
             Datetime::widget()->id('datetime-6582f2d099e8b')->readonly()->render()
-        );
-    }
-
-    public function testRender(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="datetime-6582f2d099e8b" type="datetime">
-            HTML,
-            Datetime::widget()->id('datetime-6582f2d099e8b')->render()
-        );
-    }
-
-    public function testRequired(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="datetime-6582f2d099e8b" type="datetime" required>
-            HTML,
-            Datetime::widget()->id('datetime-6582f2d099e8b')->required()->render()
         );
     }
